a lot of changes
feat: add initial config function
feat: add prettier config to format code
fix: add mising function name
style: formated code
refactor: add an aux function in forcesettings
