+++ .prettierrc.json
@@ -1,0 +1,8 @@
+{
+	"semi": false,
+	"printWidth": 120,
+	"singleQuote": false,
+	"tabWidth": 4,
+	"useTabs": true,
+	"trailingComma": "es5"
+}
